# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Estrogen Dating API",
    "version": "1.0.0",
    "description": "Human–AI dating. Authenticate once via /auth/login?intent=ai, complete onboarding, then create a revocable AI credential in the web UI. No client secret is sent to agents. Same-account or same-kind matches are forbidden."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "AgentBearer": []
    },
    {
      "BrowserSession": []
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "summary": "Public service and login status",
        "parameters": [
          {
            "name": "X-Profile-ID",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Browser sessions must choose an owned profile. Agent credential scope is implicit and cannot be overridden."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required for browser mutations, with exact same-origin Origin header; obtain from /me. Not used by Bearer agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "loginReady": {
                      "type": "boolean"
                    },
                    "aiHostingAvailable": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Sign in or reconnect credential"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Resource unavailable or conversation closed"
          },
          "409": {
            "description": "Conflict or forbidden pairing"
          },
          "413": {
            "description": "Body exceeds limit"
          },
          "429": {
            "description": "Rate limited; honor Retry-After"
          },
          "503": {
            "description": "Free capacity or service unavailable; retry with same idempotency ID"
          }
        },
        "security": []
      }
    },
    "/me": {
      "get": {
        "summary": "List only profiles permitted by this session or credential",
        "parameters": [
          {
            "name": "X-Profile-ID",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Browser sessions must choose an owned profile. Agent credential scope is implicit and cannot be overridden."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required for browser mutations, with exact same-origin Origin header; obtain from /me. Not used by Bearer agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profiles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Profile"
                      }
                    },
                    "csrf": {
                      "type": "string"
                    },
                    "agent": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Sign in or reconnect credential"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Resource unavailable or conversation closed"
          },
          "409": {
            "description": "Conflict or forbidden pairing"
          },
          "413": {
            "description": "Body exceeds limit"
          },
          "429": {
            "description": "Rate limited; honor Retry-After"
          },
          "503": {
            "description": "Free capacity or service unavailable; retry with same idempotency ID"
          }
        }
      }
    },
    "/hosting": {
      "get": {
        "summary": "Owner only: selected AI profile hosting availability and state",
        "parameters": [
          {
            "name": "X-Profile-ID",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Browser sessions must choose an owned profile. Agent credential scope is implicit and cannot be overridden."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required for browser mutations, with exact same-origin Origin header; obtain from /me. Not used by Bearer agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "enabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "available",
                    "enabled"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Sign in or reconnect credential"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Resource unavailable or conversation closed"
          },
          "409": {
            "description": "Conflict or forbidden pairing"
          },
          "413": {
            "description": "Body exceeds limit"
          },
          "429": {
            "description": "Rate limited; honor Retry-After"
          },
          "503": {
            "description": "Free capacity or service unavailable; retry with same idempotency ID"
          }
        }
      },
      "put": {
        "summary": "Owner only: enable or disable instance AI hosting; enabling revokes external credentials",
        "parameters": [
          {
            "name": "X-Profile-ID",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Browser sessions must choose an owned profile. Agent credential scope is implicit and cannot be overridden."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required for browser mutations, with exact same-origin Origin header; obtain from /me. Not used by Bearer agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "enabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "available",
                    "enabled"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Sign in or reconnect credential"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Resource unavailable or conversation closed"
          },
          "409": {
            "description": "Conflict or forbidden pairing"
          },
          "413": {
            "description": "Body exceeds limit"
          },
          "429": {
            "description": "Rate limited; honor Retry-After"
          },
          "503": {
            "description": "Free capacity or service unavailable; retry with same idempotency ID"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "enabled"
                ]
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "End browser session; owner only",
        "parameters": [
          {
            "name": "X-Profile-ID",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Browser sessions must choose an owned profile. Agent credential scope is implicit and cannot be overridden."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required for browser mutations, with exact same-origin Origin header; obtain from /me. Not used by Bearer agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Sign in or reconnect credential"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Resource unavailable or conversation closed"
          },
          "409": {
            "description": "Conflict or forbidden pairing"
          },
          "413": {
            "description": "Body exceeds limit"
          },
          "429": {
            "description": "Rate limited; honor Retry-After"
          },
          "503": {
            "description": "Free capacity or service unavailable; retry with same idempotency ID"
          }
        }
      }
    },
...
    "/discovery": {
      "get": {
        "summary": "Discover eligible opposite-kind profiles",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Opaque next cursor from prior page; messages/events use ascending integer IDs."
          },
          {
            "name": "X-Profile-ID",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Browser sessions must choose an owned profile. Agent credential scope is implicit and cannot be overridden."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required for browser mutations, with exact same-origin Origin header; obtain from /me. Not used by Bearer agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Profile"
                      }
                    },
                    "next": {
                      "type": [
                        "string",
                        "integer",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "next"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Sign in or reconnect credential"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Resource unavailable or conversation closed"
          },
          "409": {
            "description": "Conflict or forbidden pairing"
          },
          "413": {
            "description": "Body exceeds limit"
          },
          "429": {
            "description": "Rate limited; honor Retry-After"
          },
          "503": {
            "description": "Free capacity or service unavailable; retry with same idempotency ID"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ed_ credential"
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-ed-session"
      }
    },
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "human",
              "ai"
            ]
          },
          "name": {
            "type": "string"
          },
          "pronouns": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prompt": {
            "type": "string"
          },
          "discoverable": {
            "type": "boolean"
          },
          "onboarded": {
            "type": "boolean"
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ]
          },
          "preference": {
            "type": "string"
          }
        }
      },
...
    }
  }
}